+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Changed
 - Default sort on the Swagger UI is now `timestamp_cet desc`, if no sort parameter is specified `timestamp_cet asc` is used
+- Add node v16 and v18 to GitHub Action matrix
 
 ### Removed
 - Unused dependency `nock`