+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project follows [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
+
+## [1.1.0] - 2022-06-22
 ### Added
 - New `row_count` field in the response to indicate how many rows are being returned
 - 3 new query parameters: `sort`, `limit` and `offset`, they can be used for pagination
@@ -78,7 +80,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - `Fixed` for any bug fixes.
 - `Security` to invite users to upgrade in case of vulnerabilities.
 
-[Unreleased]: https://github.com/metaodi/tecdottir/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/metaodi/tecdottir/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/metaodi/tecdottir/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/metaodi/tecdottir/compare/v0.0.7...v1.0.0
 [0.0.7]: https://github.com/metaodi/tecdottir/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/metaodi/tecdottir/compare/v0.0.5...v0.0.6